Команды выбрать все

+++ xrpg/archive.h
@@ -13,7 +13,7 @@ struct archive {
 	bool				version(short major, short minor);
 	void				set(void* value, unsigned size);
 	// Array with fixed count
-	template<typename T, unsigned N> void set(T(&value)[N]) {
+	template<typename T, size_t N> void set(T(&value)[N]) {
 		for(int i = 0; i < N; i++)
 			set(value[i]);
 	};
